Update index.html
Added a new item "Languages" in the "Main features" section.

I know there is already some information about translations further down the page and it can stay there as it is, but IMO, it's nicer to provide potential users a more direct view on available languages (one-click solution). Also, the user doesn't have to leave NewPipe's website, just in order to see such information …

I'm aware, that by using/calling the Weblate's svg-image, the website contacts another server (https://github.com/TeamNewPipe/website#usage-of-website-api). I'm not a programmer, so I don't know if this can be programmed in another way. If not, maybe this could be an exception to the rule (file size is less than 50 KB)…

I don't have a chance to test the change live, so please check it.
Thank you.